WC-1391: Fix WC-1382 unit test
As it turns out, the problem was the presence of `true` and `false` in
the ZINC output, which is Not Allowed.  It's actually a WideSky bug, and
as such, the fix belongs there.

This became apparent once I made the previous commits to improve the
error reporting, it was now blindlingly obvious where the problem
existed.